+++ lib/elmas/config.rb
@@ -1,27 +1,33 @@
+# frozen_string_literal: true
+
 require "faraday"
+require "active_resource/threadsafe_attributes"
+require "logger"
 
 module Elmas
   module Config
+    include ThreadsafeAttributes
     # An array of valid keys in the options hash
-    VALID_OPTIONS_KEYS = [
-      :access_token,
-      :adapter,
-      :client_id,
-      :client_secret,
-      :connection_options,
-      :redirect_uri,
-      :response_format,
-      :user_agent,
-      :endpoint,
-      :division,
-      :base_url,
-      :refresh_token
+    VALID_OPTIONS_KEYS = %i[
+      access_token
+      adapter
+      client_id
+      client_secret
+      connection_options
+      redirect_uri
+      response_format
+      user_agent
+      endpoint
+      division
+      base_url
+      refresh_token
+      logger
     ].freeze
 
     # By default, don't set a user access token
-    DEFAULT_ACCESS_TOKEN = "".freeze
+    DEFAULT_ACCESS_TOKEN = ""
 
-    DEFAULT_REFRESH_TOKEN = "".freeze
+    DEFAULT_REFRESH_TOKEN = ""
 
     # The adapter that will be used to connect if none is set
     #
@@ -29,27 +35,27 @@ module Elmas
     DEFAULT_ADAPTER = Faraday.default_adapter
 
     # By default, client id should be set in .env
-    DEFAULT_CLIENT_ID = "".freeze
+    DEFAULT_CLIENT_ID = ""
 
     # By default, client secret should be set in .env
-    DEFAULT_CLIENT_SECRET = "".freeze
+    DEFAULT_CLIENT_SECRET = ""
 
     # By default, don't set any connection options
     DEFAULT_CONNECTION_OPTIONS = {}.freeze
 
-    DEFAULT_BASE_URL = "https://start.exactonline.nl".freeze
+    DEFAULT_BASE_URL = "https://start.exactonline.nl"
 
     # The endpoint that will be used to connect if none is set
-    DEFAULT_ENDPOINT = "api/v1".freeze
+    DEFAULT_ENDPOINT = "api/v1"
 
     # the division code you want to connect with
-    DEFAULT_DIVISION = "".freeze
+    DEFAULT_DIVISION = ""
 
     # The response format appended to the path and sent in the 'Accept' header if none is set
     #
     DEFAULT_FORMAT = :json
 
-    DEFAULT_REDIRECT_URI = "https://www.getpostman.com/oauth2/callback".freeze
+    DEFAULT_REDIRECT_URI = "https://www.getpostman.com/oauth2/callback"
 
     # By default, don't set user agent
     DEFAULT_USER_AGENT = nil
@@ -57,8 +63,10 @@ module Elmas
     # An array of valid request/response formats
     VALID_FORMATS = [:json].freeze
 
+    DEFAULT_LOGGER = ::Logger.new(STDOUT)
+
     # @private
-    attr_accessor *VALID_OPTIONS_KEYS
+    threadsafe_attribute(*VALID_OPTIONS_KEYS)
 
     # When this module is extended, set all configuration options to their default values
     def self.extended(base)
@@ -91,6 +99,7 @@ module Elmas
       self.response_format    = DEFAULT_FORMAT
       self.user_agent         = DEFAULT_USER_AGENT
       self.refresh_token      = DEFAULT_REFRESH_TOKEN
+      self.logger             = DEFAULT_LOGGER
     end
   end
 end