+++ lib/elmas/oauth.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require "mechanize"
 require "uri"
 require "json"
@@ -7,6 +9,7 @@ require File.expand_path("../response", __FILE__)
 
 # from https://developers.exactonline.com/#Example retrieve access token.html
 module Elmas
+  # rubocop:disable Metrics/ModuleLength
   module OAuth
     def authorize(user_name, password, options = {})
       agent = Mechanize.new
@@ -16,6 +19,15 @@ module Elmas
 
       code = URI.unescape(agent.page.uri.query.split("=").last)
       OauthResponse.new(get_access_token(code))
+    end
+
+    def refresh_authorization
+      OauthResponse.new(get_refresh_token(refresh_token)).tap do |response|
+        Elmas.configure do |config|
+          config.access_token = response.access_token
+          config.refresh_token = response.refresh_token
+        end
+      end
     end
 
     def authorized?
@@ -32,13 +44,10 @@ module Elmas
 
     def auto_authorize
       Elmas.configure do |config|
+        config.redirect_uri = ENV["REDIRECT_URI"]
         config.client_id = ENV["CLIENT_ID"]
         config.client_secret = ENV["CLIENT_SECRET"]
-      end
-      Elmas.configure do |config|
         config.access_token = Elmas.authorize(ENV["EXACT_USER_NAME"], ENV["EXACT_PASSWORD"]).access_token
-      end
-      Elmas.configure do |config|
         config.division = Elmas.authorize_division
       end
     end
@@ -67,6 +76,23 @@ module Elmas
       end
     end
 
+    # Return an access token from authorization via refresh token
+    def get_refresh_token(refresh_token)
+      conn = Faraday.new(url: config[:base_url]) do |faraday|
+        faraday.request :url_encoded
+        faraday.adapter Faraday.default_adapter
+      end
+
+      params = refresh_access_token_params(refresh_token)
+
+      conn.post do |req|
+        req.url "/api/oauth2/token"
+        req.body = params
+        req.headers["Accept"] = "application/json"
+        req.headers["Content-Type"] = "application/x-www-form-urlencoded"
+      end
+    end
+
     private
 
     def login(agent, user_name, password, options)
@@ -81,6 +107,7 @@ module Elmas
 
     def allow_access(agent)
       return if agent.page.uri.to_s.include?("getpostman")
+      return if agent.page.uri.to_s.include?(redirect_uri)
       form = agent.page.form_with(id: "PublicOAuth2Form")
       button = form.button_with(id: "AllowButton")
       agent.submit(form, button)
@@ -99,6 +126,15 @@ module Elmas
         grant_type: "authorization_code",
         code: code,
         redirect_uri: redirect_uri
+      }
+    end
+
+    def refresh_access_token_params(code)
+      {
+        client_id: client_id,
+        client_secret: client_secret,
+        grant_type: "refresh_token",
+        refresh_token: code
       }
     end
   end