+++ lib/elmas/resources/mailbox.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Elmas
   class Mailbox
     include Elmas::Resource
@@ -7,14 +9,14 @@ module Elmas
     end
 
     def other_attributes
-      [
-        :account, :description, :for_division, :publish,
-        :type, :valid_from, :valid_to
+      %i[
+        account description for_division publish
+        type valid_from valid_to
       ]
     end
 
     def mandatory_attributes
-      [:mailbox]
+      %i[mailbox]
     end
   end
 end