+++ lib/elmas/resources/printed_sales_invoice.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Elmas
   # We can use the PrintedSalesInvoice to change the status of SalesInvoices from
   # Open to 'Verwerkt' while at the same time sending a PDF of the invoice to the
@@ -9,7 +11,7 @@ module Elmas
     include Elmas::Resource
 
     def valid_actions
-      [:post]
+      %i[post]
     end
 
     def base_path
@@ -17,18 +19,18 @@ module Elmas
     end
 
     def mandatory_attributes
-      [:invoice_ID]
+      %i[invoice_ID]
     end
 
-    # https://start.exactonline.nl/docs/HlpRestAPIResourcesDetails.aspx?name=SalesInvoicePrintedSalesInvoices
+    # https//start.exactonline.nl/docs/HlpRestAPIResourcesDetails.aspx?name=SalesInvoicePrintedSalesInvoices
     def other_attributes
-      [
-        :division, :document, :document_creation_error, :document_creation_success,
-        :document_layout, :email_creation_error, :email_creation_success, :email_layout,
-        :extra_text, :invoice_date, :postbox_message_creation_error,
-        :postbox_message_creation_success, :postbox_sender, :reporting_period,
-        :reporting_year, :send_email_to_customer, :send_invoice_to_customer_postbox,
-        :send_output_based_on_account
+      %i[
+        division document document_creation_error document_creation_success
+        document_layout email_creation_error email_creation_success email_layout
+        extra_text invoice_date postbox_message_creation_error
+        postbox_message_creation_success postbox_sender reporting_period
+        reporting_year send_email_to_customer send_invoice_to_customer_postbox
+        send_output_based_on_account
       ]
     end
   end