+++ lib/elmas/resources/purchase_entry.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Elmas
   class PurchaseEntry
     include Elmas::Resource
@@ -7,15 +9,12 @@ module Elmas
     end
 
     def mandatory_attributes
-      [:journal, :purchase_entry_lines, :supplier]
+      %i[journal purchase_entry_lines supplier]
     end
 
     def other_attributes
-      [
-        :currency, :batch_number, :description, :document, :due_date, :entry_date,
-        :entry_number, :external_link_reference, :invoice_number, :rate, :reporting_period,
-        :reporting_year, :reversal, :VAT_amount_FC, :your_ref
-      ]
+      %i[currency batch_number description document due_date entry_date entry_number external_link_reference invoice_number rate reporting_period
+         reporting_year reversal VAT_amount_FC your_ref]
     end
   end
 end