+++ lib/elmas/resources/sales_invoice.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Elmas
   class SalesInvoice
     # An sales_invoice usually has multiple sales_invoice lines
@@ -10,15 +12,12 @@ module Elmas
     end
 
     def mandatory_attributes
-      [:journal, :ordered_by]
+      [:ordered_by]
     end
 
     def other_attributes
       SHARED_SALES_ATTRIBUTES.inject(
-        [
-          :sales_invoice_lines, :due_date, :sales_person,
-          :starter_sales_invoice_status, :type
-        ],
+        %i[sales_invoice_lines due_date sales_person starter_sales_invoice_status type],
         :<<
       )
     end