+++ lib/elmas/resources/transaction_line.rb
@@ -1,9 +1,11 @@
+# frozen_string_literal: true
+
 module Elmas
   class TransactionLine
     include Elmas::Resource
 
     def valid_actions
-      [:get]
+      %i[get]
     end
 
     def base_path
@@ -15,7 +17,7 @@ module Elmas
     end
 
     def other_attributes
-      [:asset_code]
+      %i[asset_code]
     end
   end
 end