+++ Changelog.md
@@ -1,3 +1,10 @@
+## 2.4.2
+- Add thread safety
+
+## 2.4.1
+
+- Update gem versions of mechanize and faraday
+
 ## 2.4.0
 
 - Add security measures for constantizing Exact Resources thanks to @confiks